adding support for bfloat16 models

+++ bernn/dl/models/pytorch/aedann.py
@@ -107,12 +107,12 @@ class Classifier(nn.Module):
         x = self.net(x)
         if self.use_softmax:
             x = F.softmax(x, dim=1)
-        return x.detach().cpu().numpy()
+        return x.detach().float().cpu().numpy()
 
     def predict(self, x: torch.Tensor) -> np.ndarray:
         x = self.net(x)
         x = x.argmax(1)
-        return x.detach().cpu().numpy()
+        return x.detach().float().cpu().numpy()
 
 
 class Classifier2(nn.Module):
@@ -146,14 +146,14 @@ class Classifier2(nn.Module):
             #     nn.init.constant_(m.bias, 0.125)
 
     def predict_proba(self, x: torch.Tensor) -> np.ndarray:
-        x = self.linear2(x).detach().cpu().numpy()
+        x = self.linear2(x).detach().float().cpu().numpy()
         if self.use_softmax:
             x = F.softmax(x, dim=1)
 
         return x
 
     def predict(self, x: torch.Tensor) -> np.ndarray:
-        return self.linear2(x).argmax(1).detach().cpu().numpy()
+        return self.linear2(x).argmax(1).detach().float().cpu().numpy()
 
 
 class Encoder(nn.Module):
@@ -516,10 +516,10 @@ class SHAPAutoEncoder2(nn.Module):
                 nn.init.constant_(m.bias, 0.125)
 
     def predict_proba(self, x: torch.Tensor) -> np.ndarray:
-        return self.classifier(x).detach().cpu().numpy()
+        return self.classifier(x).detach().float().cpu().numpy()
 
     def predict(self, x: torch.Tensor) -> np.ndarray:
-        return self.classifier(x).argmax(1).detach().cpu().numpy()
+        return self.classifier(x).argmax(1).detach().float().cpu().numpy()
 
     def _kld(self, z: torch.Tensor, q_param: Tuple[torch.Tensor, torch.Tensor],
              h_last: Optional[torch.Tensor] = None,
@@ -732,7 +732,7 @@ class SHAPAutoEncoder3(nn.Module):
         Returns:
             np.ndarray: Probability predictions
         """
-        return self.classifier(x).detach().cpu().numpy()
+        return self.classifier(x).detach().float().cpu().numpy()
 
     def predict(self, x: torch.Tensor) -> np.ndarray:
         """Get class predictions.
@@ -743,7 +743,7 @@ class SHAPAutoEncoder3(nn.Module):
         Returns:
             np.ndarray: Class predictions
         """
-        return self.classifier(x).argmax(1).detach().cpu().numpy()
+        return self.classifier(x).argmax(1).detach().float().cpu().numpy()
 
     def _kld(self, z: torch.Tensor, q_param: Tuple[torch.Tensor, torch.Tensor],
              h_last: Optional[torch.Tensor] = None,
@@ -947,10 +947,10 @@ class AutoEncoder2(nn.Module):
             #     nn.init.constant_(m.bias, 0.125)
 
     def predict_proba(self, x: torch.Tensor) -> np.ndarray:
-        return self.classifier(x).detach().cpu().numpy()
+        return self.classifier(x).detach().float().cpu().numpy()
 
     def predict(self, x: torch.Tensor) -> np.ndarray:
-        return self.classifier(x).argmax(1).detach().cpu().numpy()
+        return self.classifier(x).argmax(1).detach().float().cpu().numpy()
 
     def _kld(self, z: torch.Tensor, q_param: Tuple[torch.Tensor, torch.Tensor],
              h_last: Optional[torch.Tensor] = None,
@@ -1142,11 +1142,11 @@ class AutoEncoder3(nn.Module):
 
     def predict_proba(self, inputs: torch.Tensor) -> np.ndarray:
         x = self.enc(inputs)
-        return self.classifier(x).detach().cpu().float().numpy()
+        return self.classifier(x).detach().float().cpu().float().numpy()
 
     def predict(self, inputs: torch.Tensor) -> np.ndarray:
         x = self.enc(inputs)
-        return self.classifier(x).argmax(1).detach().cpu().float().numpy()
+        return self.classifier(x).argmax(1).detach().float().cpu().float().numpy()
 
     def _kld(self, z: torch.Tensor, q_param: Tuple[torch.Tensor, torch.Tensor],
              h_last: Optional[torch.Tensor] = None,
