+++ bernn/dl/models/pytorch/aekandann.py
@@ -114,10 +114,10 @@ class Classifier(nn.Module):
                     m.bias.data.zero_()
 
     def predict_proba(self, x):
-        return self.linear2(x).detach().cpu().numpy()
+        return self.linear2(x).detach().float().cpu().numpy()
 
     def predict(self, x):
-        return self.linear2(x).argmax(1).detach().cpu().numpy()
+        return self.linear2(x).argmax(1).detach().float().cpu().numpy()
 
 
 class Classifier2(nn.Module):
@@ -150,10 +150,10 @@ class Classifier2(nn.Module):
             #     nn.init.constant_(m.bias, 0.125)
 
     def predict_proba(self, x):
-        return self.linear2(x).detach().cpu().numpy()
+        return self.linear2(x).detach().float().cpu().numpy()
 
     def predict(self, x):
-        return self.linear2(x).argmax(1).detach().cpu().numpy()
+        return self.linear2(x).argmax(1).detach().float().cpu().numpy()
 
 
 class Classifier3(nn.Module):
@@ -193,10 +193,10 @@ class Classifier3(nn.Module):
             #     nn.init.constant_(m.bias, 0.125)
 
     def predict_proba(self, x):
-        return self.linear2(x).detach().cpu().numpy()
+        return self.linear2(x).detach().float().cpu().numpy()
 
     def predict(self, x):
-        return self.linear2(x).argmax(1).detach().cpu().numpy()
+        return self.linear2(x).argmax(1).detach().float().cpu().numpy()
 
 
 class Encoder(nn.Module):
@@ -402,10 +402,10 @@ class SHAPKANAutoEncoder2(nn.Module):
                 nn.init.constant_(m.bias, 0.125)
 
     def predict_proba(self, x):
-        return self.classifier(x).detach().cpu().numpy()
+        return self.classifier(x).detach().float().cpu().numpy()
 
     def predict(self, x):
-        return self.classifier(x).argmax(1).detach().cpu().numpy()
+        return self.classifier(x).argmax(1).detach().float().cpu().numpy()
 
     def _kld(self, z, q_param, h_last=None, p_param=None):
         if len(z.shape) == 1:
@@ -564,10 +564,10 @@ class KANAutoEncoder2(nn.Module):
             #     nn.init.constant_(m.bias, 0.125)
 
     def predict_proba(self, x):
-        return self.classifier(x).detach().cpu().numpy()
+        return self.classifier(x).detach().float().cpu().numpy()
 
     def predict(self, x):
-        return self.classifier(x).argmax(1).detach().cpu().numpy()
+        return self.classifier(x).argmax(1).detach().float().cpu().numpy()
 
     def _kld(self, z, q_param, h_last=None, p_param=None):
         if len(z.shape) == 1:
@@ -813,10 +813,10 @@ class SHAPKANAutoEncoder3(nn.Module):
                 nn.init.constant_(m.bias, 0.125)
 
     def predict_proba(self, x):
-        return self.classifier(x).detach().cpu().numpy()
+        return self.classifier(x).detach().float().cpu().numpy()
 
     def predict(self, x):
-        return self.classifier(x).argmax(1).detach().cpu().numpy()
+        return self.classifier(x).argmax(1).detach().float().cpu().numpy()
 
     def _kld(self, z, q_param, h_last=None, p_param=None):
         if len(z.shape) == 1:
@@ -976,10 +976,10 @@ class KANAutoEncoder3(nn.Module):
                 m.reset_parameters()
 
     def predict_proba(self, x):
-        return self.classifier(x).detach().cpu().numpy()
+        return self.classifier(x).detach().float().cpu().numpy()
 
     def predict(self, x):
-        return self.classifier(x).argmax(1).detach().cpu().numpy()
+        return self.classifier(x).argmax(1).detach().float().cpu().numpy()
 
     def _kld(self, z, q_param, h_last=None, p_param=None):
         if len(z.shape) == 1:
