+++ bernn/dl/models/pytorch/utils/dataset.py
@@ -503,7 +503,7 @@ def get_loaders(data, random_recs, samples_weights, triplet_dloss, ae=None, clas
             enc, rec, _, kld = ae(input, domain, sampling=False)
             preds = ae.classifier(enc)
             domain_preds = ae.dann_discriminator(enc)
-            valid_cats += [preds.detach().cpu().numpy().argmax(1)]
+            valid_cats += [preds.detach().float().cpu().numpy().argmax(1)]
             valid_names += [names]
         for i, batch in enumerate(loaders['test']):
             # optimizer_ae.zero_grad()
@@ -517,7 +517,7 @@ def get_loaders(data, random_recs, samples_weights, triplet_dloss, ae=None, clas
             domain_preds = ae.dann_discriminator(enc)
             # else:
             #     preds = classifier(enc)
-            test_cats += [preds.detach().cpu().numpy().argmax(1)]
+            test_cats += [preds.detach().float().cpu().numpy().argmax(1)]
             test_names += [names]
 
         valid_set2 = MSDataset3(data['inputs']['valid'], valid_names, np.concatenate(valid_cats),
@@ -704,7 +704,7 @@ def get_images_loaders(data, random_recs, samples_weights, triplet_dloss, ae=Non
             enc, rec, _, kld = ae(input, domain, sampling=False)
             preds = ae.classifier(enc)
             domain_preds = ae.dann_discriminator(enc)
-            valid_cats += [preds.detach().cpu().numpy().argmax(1)]
+            valid_cats += [preds.detach().float().cpu().numpy().argmax(1)]
             valid_names += [names]
         for i, batch in enumerate(loaders['test']):
             # optimizer_ae.zero_grad()
@@ -718,7 +718,7 @@ def get_images_loaders(data, random_recs, samples_weights, triplet_dloss, ae=Non
             domain_preds = ae.dann_discriminator(enc)
             # else:
             #     preds = classifier(enc)
-            test_cats += [preds.detach().cpu().numpy().argmax(1)]
+            test_cats += [preds.detach().float().cpu().numpy().argmax(1)]
             test_names += [names]
 
         valid_set2 = MSDataset3(data['inputs']['valid'], valid_names, np.concatenate(valid_cats),
@@ -911,7 +911,7 @@ def get_images_loaders_no_pool(data, random_recs, samples_weights, triplet_dloss
             enc, rec, _, kld = ae(input, domain, sampling=False)
             preds = ae.classifier(enc)
             domain_preds = ae.dann_discriminator(enc)
-            valid_cats += [preds.detach().cpu().numpy().argmax(1)]
+            valid_cats += [preds.detach().float().cpu().numpy().argmax(1)]
             valid_names += [names]
         for i, batch in enumerate(loaders['test']):
             # optimizer_ae.zero_grad()
@@ -925,7 +925,7 @@ def get_images_loaders_no_pool(data, random_recs, samples_weights, triplet_dloss
             domain_preds = ae.dann_discriminator(enc)
             # else:
             #     preds = classifier(enc)
-            test_cats += [preds.detach().cpu().numpy().argmax(1)]
+            test_cats += [preds.detach().float().cpu().numpy().argmax(1)]
             test_names += [names]
 
         valid_set2 = MSDataset5(data['inputs']['valid'], valid_names, np.concatenate(valid_cats),
@@ -1063,7 +1063,7 @@ def get_loaders_no_pool(data, random_recs, samples_weights, triplet_dloss, ae=No
             enc, rec, _, kld = ae(input, domain, sampling=False)
             preds = ae.classifier(enc)
             domain_preds = ae.dann_discriminator(enc)
-            valid_cats += [preds.detach().cpu().numpy().argmax(1)]
+            valid_cats += [preds.detach().float().cpu().numpy().argmax(1)]
             valid_names += [names]
         for i, batch in enumerate(loaders['test']):
             # optimizer_ae.zero_grad()
@@ -1077,7 +1077,7 @@ def get_loaders_no_pool(data, random_recs, samples_weights, triplet_dloss, ae=No
             domain_preds = ae.dann_discriminator(enc)
             # else:
             #     preds = classifier(enc)
-            test_cats += [preds.detach().cpu().numpy().argmax(1)]
+            test_cats += [preds.detach().float().cpu().numpy().argmax(1)]
             test_names += [names]
 
         valid_set2 = MSDataset3(data['inputs']['valid'], valid_names, np.concatenate(valid_cats),
@@ -1189,10 +1189,10 @@ class MSCSV:
         if self.resize:
             try:
                 mat_data = transforms.Resize((self.new_size, self.new_size))(
-                torch.Tensor(mat_data.values).unsqueeze(0)).squeeze().detach().cpu().numpy()
+                torch.Tensor(mat_data.values).unsqueeze(0)).squeeze().detach().float().cpu().numpy()
             except:
                 mat_data = transforms.Resize((self.new_size, self.new_size))(
-                torch.Tensor(mat_data).unsqueeze(0)).squeeze().detach().cpu().numpy()
+                torch.Tensor(mat_data).unsqueeze(0)).squeeze().detach().float().cpu().numpy()
 
         return mat_data.astype('float'), label, batch, plate, fname.split('.csv')[0]
 
@@ -1269,10 +1269,10 @@ class MS2CSV:
             if self.resize:
                 try:
                     mat_data = transforms.Resize((self.new_size, self.new_size))(
-                    torch.Tensor(mat_data.values).unsqueeze(0)).squeeze().detach().cpu().numpy()
+                    torch.Tensor(mat_data.values).unsqueeze(0)).squeeze().detach().float().cpu().numpy()
                 except:
                     mat_data = transforms.Resize((self.new_size, self.new_size))(
-                    torch.Tensor(mat_data).unsqueeze(0)).squeeze().detach().cpu().numpy()
+                    torch.Tensor(mat_data).unsqueeze(0)).squeeze().detach().float().cpu().numpy()
             mat_datas += [mat_data]
         # try:
         mat_data = np.stack(mat_datas, 0)
@@ -1360,7 +1360,7 @@ class MSCSV2:
 
         if self.resize:
             mat_data = transforms.Resize((299, 299))(
-                torch.Tensor(mat_data).unsqueeze(0)).squeeze().detach().cpu().numpy()
+                torch.Tensor(mat_data).unsqueeze(0)).squeeze().detach().float().cpu().numpy()
 
         return mat_data.astype('float'), label, low, batch, fname
