+++ bernn/dl/train/aecnndann.py
@@ -113,10 +113,10 @@ class Classifier(nn.Module):
             #     nn.init.constant_(m.bias, 0.125)
 
     def predict_proba(self, x):
-        return self.linear2(x).detach().cpu().numpy()
+        return self.linear2(x).detach().float().cpu().numpy()
 
     def predict(self, x):
-        return self.linear2(x).argmax(1).detach().cpu().numpy()
+        return self.linear2(x).argmax(1).detach().float().cpu().numpy()
 
 
 class Classifier2(nn.Module):
@@ -149,10 +149,10 @@ class Classifier2(nn.Module):
             #     nn.init.constant_(m.bias, 0.125)
 
     def predict_proba(self, x):
-        return self.linear2(x).detach().cpu().numpy()
+        return self.linear2(x).detach().float().cpu().numpy()
 
     def predict(self, x):
-        return self.linear2(x).argmax(1).detach().cpu().numpy()
+        return self.linear2(x).argmax(1).detach().float().cpu().numpy()
 
 
 class Classifier3(nn.Module):
@@ -192,10 +192,10 @@ class Classifier3(nn.Module):
             #     nn.init.constant_(m.bias, 0.125)
 
     def predict_proba(self, x):
-        return self.linear2(x).detach().cpu().numpy()
+        return self.linear2(x).detach().float().cpu().numpy()
 
     def predict(self, x):
-        return self.linear2(x).argmax(1).detach().cpu().numpy()
+        return self.linear2(x).argmax(1).detach().float().cpu().numpy()
 
 
 class ConvEncoder(nn.Module):
@@ -400,10 +400,10 @@ class SHAPAutoEncoderCNN(nn.Module):
                 nn.init.constant_(m.bias, 0.125)
 
     def predict_proba(self, x):
-        return self.classifier(x).detach().cpu().numpy()
+        return self.classifier(x).detach().float().cpu().numpy()
 
     def predict(self, x):
-        return self.classifier(x).argmax(1).detach().cpu().numpy()
+        return self.classifier(x).argmax(1).detach().float().cpu().numpy()
 
     def _kld(self, z, q_param, h_last=None, p_param=None):
         if len(z.shape) == 1:
@@ -557,10 +557,10 @@ class AutoEncoderCNN(nn.Module):
             #     nn.init.constant_(m.bias, 0.125)
 
     def predict_proba(self, x):
-        return self.classifier(x).detach().cpu().numpy()
+        return self.classifier(x).detach().float().cpu().numpy()
 
     def predict(self, x):
-        return self.classifier(x).argmax(1).detach().cpu().numpy()
+        return self.classifier(x).argmax(1).detach().float().cpu().numpy()
 
     def _kld(self, z, q_param, h_last=None, p_param=None):
         if len(z.shape) == 1:
